pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sienge-ecbiesek-mcp"
version = "1.2.1"
description = "🏗️ Model Context Protocol (MCP) server for Sienge API integration - Brazilian construction ERP system. Connect Claude AI to Sienge with 50+ powerful tools for comprehensive business management including financials, projects, and operations."
authors = [
    {name = "ECBIESEK", email = "ti@ecbiesek.com"}
]
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = [
    "sienge", 
    "mcp", 
    "model-context-protocol", 
    "claude", 
    "api", 
    "construction", 
    "erp", 
    "brazil", 
    "ecbiesek",
    "fastmcp",
    "ai-integration"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Environment :: Console",
    "Framework :: FastAPI"
]

dependencies = [
    "fastmcp>=2.12.3",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "build>=0.10.0",
    "twine>=4.0.0",
    "toml>=0.10.2",
]

[project.urls]
Homepage = "https://github.com/INOTECH-ecbiesek/Sienge-MCP"
Documentation = "https://github.com/INOTECH-ecbiesek/Sienge-MCP#readme"
Repository = "https://github.com/INOTECH-ecbiesek/Sienge-MCP.git"
Issues = "https://github.com/INOTECH-ecbiesek/Sienge-MCP/issues"
"Bug Reports" = "https://github.com/INOTECH-ecbiesek/Sienge-MCP/issues"
"Source Code" = "https://github.com/INOTECH-ecbiesek/Sienge-MCP"
"PyPI" = "https://pypi.org/project/sienge-ecbiesek-mcp/"

[project.scripts]
sienge-ecbiesek-mcp = "sienge_mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sienge_mcp = ["*.md", "*.txt", "*.json"]
